modified:   app.js  	modified:   package-lock.json  	modified:   package.json  	modified:   routes/authorization.js  	modified:   routes/routes.js  	modified:   service/authorization.js  	new file:   service/functions.js
