+++ app.js
@@ -22,18 +22,20 @@ const connectDatabase = async () => {
 
 connectDatabase();
 const app = express();
-const contactsRouter = require("./routes/contacts");
-const authRouter = require("./routes/authorization");
+// const contactsRouter = require("./routes/contacts");
+// const authRouter = require("./routes/authorization");
 const formatsLogger = app.get("env") === "development" ? "dev" : "short";
-
+const router = require("./routes/routes");
 
 app.use(morgan(formatsLogger))
 app.use(cors())
 app.use(express.json())
 
-app.use('/api/contacts', contactsRouter)
-app.use('/api/users', authRouter)
-  
+// app.use('/api/contacts', contactsRouter)
+// app.use('/users', authRouter)
+
+app.use(router); 
+
 app.use((req, res) => {
   res.status(404).json({ message: "Not found" })
 });
