+++ routes/routes.js
@@ -1,8 +1,14 @@
-// const express = require("express");
-// const contactRouter = require("./contacts");
+const express = require("express");
+const contactsRouter = require("./contacts");
+const authRouter = require("./authorization");
  
-// const router = express.Router();
+const router = express.Router();
 
 // router.use("/api/contacts", contactRouter);
 
 // module.exports = router;
+
+router.use("/api/contacts", contactsRouter);
+router.use("/users", authRouter);
+
+module.exports = router;
